Refactor components for questionnnaire-builder
