--- src/components/questionnaire-builder/QuestionList/FrameworkQuestionForm/FrameworkAttributeInput/styles.scss
@@ -1,14 +1,0 @@
-.framework-attribute-input {
-    display: flex;
-    padding: 0 calc(var(--spacing-medium) - var(--spacing-small));
-
-    .input {
-        margin: var(--spacing-small);
-    }
-
-    .sector-input-container,
-    .dimension-input-container {
-        display: flex;
-        flex-grow: 1;
-    }
-}
