--- src/components/questionnaire-builder/QuestionList/styles.scss
@@ -1,31 +1,0 @@
-.questions {
-    display: flex;
-    flex-direction: column;
-
-    .header {
-        display: flex;
-        align-items: center;
-        flex-shrink: 0;
-
-        .heading {
-            flex-grow: 1;
-            padding: var(--spacing-medium);
-        }
-
-        .actions {
-            flex-shrink: 0;
-            padding: 0 var(--spacing-medium);
-        }
-    }
-
-
-    .content {
-        flex-grow: 1;
-        overflow: auto;
-    }
-}
-
-.question {
-    margin: var(--spacing-small) var(--spacing-medium);
-    background-color: var(--color-foreground);
-}
