+++ src/components/questionnaire-builder/QuestionModal/ResponseInput/styles.scss
