+++ src/components/questionnaire-builder/QuestionModal/styles.scss
