+++ src/components/questionnaire-builder/QuestionnaireModal/styles.scss
