+++ src/declarations/png.d.ts
@@ -1,0 +1,4 @@
+declare module '*.png' {
+    const value: any;
+    export = value;
+}
