+++ src/views/FrameworkQuestions/styles.scss
@@ -17,9 +17,39 @@
         display: flex;
 
         .question-list {
+            display: flex;
+            flex-direction: column;
             flex-grow: 1;
             overflow: auto;
+
+            .header {
+                display: flex;
+                align-items: center;
+                flex-shrink: 0;
+
+                .heading {
+                    flex-grow: 1;
+                    padding: var(--spacing-medium);
+                }
+
+                .actions {
+                    flex-shrink: 0;
+                    padding: 0 var(--spacing-medium);
+                }
+            }
+
+
+            .content {
+                flex-grow: 1;
+                overflow: auto;
+            }
         }
+
+        .question {
+            margin: var(--spacing-small) var(--spacing-medium);
+            background-color: var(--color-foreground);
+        }
+
 
         .right-panel {
             display: flex;
