+++ src/views/Project/Details/Frameworks/FrameworkDetail/Preview/QuestionList/styles.scss
@@ -1,0 +1,9 @@
+.content {
+    flex-grow: 1;
+    overflow: auto;
+
+    .question {
+        margin: var(--spacing-small) var(--spacing-medium);
+        background-color: var(--color-foreground);
+    }
+}
